Move __version__ so that setup.py has no dependencies
