+++ doc/conf.py
@@ -13,7 +13,7 @@
 
 import sys, os
 
-from sphinxcontrib.doxylink.doxylink import __version__
+from sphinxcontrib.doxylink import __version__
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
